Changed setupTags.sh to groups minor builds
If the first minor works, assume they all work.
And if it fails, assume they all will.

+++ setupTags.sh
@@ -10,37 +10,51 @@ VERSIONS=$(curl -sS https://mirrors.edge.kernel.org/pub/software/scm/git/ | sed 
 LATEST_VERSION="$(echo "$VERSIONS" | tail -1)"
 
 LATEST_COMMIT=""
+LAST_WORKING_MINOR="1.8"
+LAST_BROKEN_MINOR="X.X"
 
 function doVersion {
     local version=$1
-    for dockerfile in Dockerfile-*
-    do
-        echo "Doing version: $version ($dockerfile)"
-        sed -i "s/ARG VERSION=.*/ARG VERSION=$version/" $dockerfile
-        set +e
-        docker build . --file $dockerfile --tag git_tmp
-        exit_status=$?
-        if [[ "$exit_status" != "0" ]]
-        then
-            echo "Buid failed"
-            return
-        fi
-        docker run --rm -it git_tmp --version | grep "$version"
-        if [[ "$exit_status" != "0" ]]
-        then
-            echo "Buid corrupt somehow"
-            return
-        fi
-        set -e
-        echo "Build success"
-    done
-    echo "All builds successfull. Pushing tags"
+    sed -i "s/ARG VERSION=.*/ARG VERSION=$version/" Dockerfile-*
+    if [[ "$version" =~ ^$LAST_WORKING_MINOR ]]
+    then
+        echo "Assuming it will work as minor worked: $LAST_WORKING_MINOR"
+    elif [[ "$version" =~ ^$LAST_BROKEN_MINOR ]]
+    then
+        echo "Assuming it will NOT work as minor did not: $LAST_BROKEN_MINOR"
+    else
+        for dockerfile in Dockerfile-*
+        do
+            echo "Doing version: $version ($dockerfile)"
+            set +e
+            docker build . --file $dockerfile --tag git_tmp
+            exit_status=$?
+            if [[ "$exit_status" != "0" ]]
+            then
+                echo "Buid failed"
+                LAST_BROKEN_MINOR="$(echo $version | cut -d. -f1,2)"
+                return
+            fi
+            docker run --rm -it git_tmp --version | grep "$version"
+            if [[ "$exit_status" != "0" ]]
+            then
+                echo "Buid corrupt somehow"
+                LAST_BROKEN_MINOR="$(echo $version | cut -d. -f1,2)"
+                return
+            fi
+            set -e
+            echo "Build success"
+        done
+        echo "All builds successfull"
+    fi
+    echo "Pushing tags"
     git reset origin/master
     git add -- Dockerfile-*
     git commit --allow-empty -m "AUTOMATIC COMMIT FOR $version" >/dev/null 2>&1
     git tag -f $version
     git push --tags --force
     LATEST_COMMIT="$version"
+    LAST_WORKING_MINOR="$(echo $version | cut -d. -f1,2)"
 }
 
 git checkout origin/master
@@ -55,6 +69,7 @@ do
     then
         echo "Skipping version: $version"
         LATEST_COMMIT="$version"
+        LAST_WORKING_MINOR="$(echo $version | cut -d. -f1,2)"
         continue
     else
         doVersion $version
